Add MPS and STS computations for intrapop
Also correct some MPS and STS related things (like only 3 neurons were
used to compute this indexes instead of the whole pop).

The MPS and STS results are now stored in dictionnaries. Each
dictionnary has keys for each subpop number, and a 'whole' key for the
index on the whole population.