Add global resampling for simulation records
In order to use less memory (both RAM during the simulation, and disk
space when saving data to files) the signals can now be recorded at a
different rate than the simulation timestep.

Previous resampling for data recording is now gone because the signals
already have a lower resolution when they are recorded.

The parameter to control the signal record timestep is
['Common']['resample_dt'].