Fix fftmax mean computation
Previously, the mean of the FFTmax index was defined as the FFTmax for
the mean signal. However, if the N populations were not in phase, the
resulting FFT max was N times higher.

The mean of the FFTmax is now defined as mean of all the subpopulation
FFTmax, thus avoiding to create artificial subfrequency.